refactor: Separate monolith and add a "Player" class
Change
 - Separate functions into separate .py files based on their purpose
 - Create a basic player class
 - Add a main file to bind multiple
 - Move responses to user input into a json file

The original file is a hodgepodge of working functions relating
to the leaderboard, the player, and the game itself. I didn't like that;
So the logic for each aforementioned portion was snapped into their own
files. A driver file acts as the main function to unite and run the
game.

In refactoring into separate files, I decided to add a dash of OO
to the game and cobble together a Player class. I toyed with the idea of
shelving the player objects, but decided against it. I'm not familiar
enough with OOP and I already have a nerdy private projects I'm working on
aimed at strengthening that muscle. Ultimately, the player class is far
from polished. It should serve a sufficient platform to build from
should the desire ever arise.

Finally, the programmed responses to user input where moved to a JSON
file. The initial goal for the project was to create something with a
little more life than the traditional number guessing game. The issue
was the noise produced by storing the responses internally. Outsourcing
the responsibility to a JSON file keeps the dictionary structure in tact
with minimal noise added to the code.